refactor: extract pluggable TLS backend abstraction
Introduce a backend package with a Backend interface and SSLOpts, and route
OpenSSLDialer through it instead of linking OpenSSL directly. The cgo OpenSSL
engine becomes the first Backend implementation and lives in its own
import-gated sub-package (backend/openssl, //go:build cgo, openssl.New).

OpenSSLDialer.Backend must now be set explicitly (e.g. openssl.New()); Dial
returns an error when it is nil. The root tlsdialer package imports no TLS
engine, so importing it never pulls in cgo — a program opts into OpenSSL/cgo
only by importing backend/openssl.

ttConn is now engine-agnostic and owns the backend's net.Conn. All external
tests move into the integration package behind a backend-parameterised harness.

README and CHANGELOG document the backend abstraction and the cgo logic
(including the CGO_ENABLED=0 compilation-error example).
